+++ Backend/routes/movies.js
@@ -5,16 +5,7 @@ var multer = require('multer');
 
 var movieController = require('../controllers/movies');
 
-//Specifying the storage path for the movie
-var storage = multer.diskStorage({
-    destination: function(req, file, cb){
-        cb(null, './uploads/');
-    },
-    filename: function(req, file, cb){
-        cb(null, file.originalname);
-    }
-})
-var upload = multer({storage: storage});
+var upload = require('../movieUpload')
 
 
 router.post('/', upload.single('movieFile'), movieController.addMovie);
